fixed: decreasing speed below 1, switched from long to float

+++ src/afk/game/AFKGameCoordinator.java
@@ -67,7 +67,7 @@ public class AFKGameCoordinator implements GameCoordinator {
     }
 
     @Override
-        public long getGameSpeed(){
+        public float getGameSpeed(){
         return gameEngine.getSpeed();
     }
